convert to workspace
