+++ cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "rust_rest_cli"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[[bin]]
+name = "db"
+path = "src/bin/db.rs"
+
+[[bin]]
+name = "generate"
+path = "src/bin/generate.rs"
+
+[dependencies]
+axum-on-rails = { path = "../axum-on-rails" }
+tokio = { version = "1.33", features = ["full"] }
