+++ cli/src/bin/db.rs
