+++ cli/src/bin/generate.rs
