+++ config/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rust_rest_config"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
