+++ db/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rust_rest_db"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] } 
+uuid = { version = "1.5", features = ["serde"] }
