+++ db/src/entities/task.rs
@@ -1,10 +1,8 @@
-#[cfg(test)]
 use serde::Deserialize;
 use serde::Serialize;
 use uuid::Uuid;
 
-#[derive(Serialize, Debug)]
-#[cfg_attr(test, derive(Deserialize))]
+#[derive(Serialize, Debug, Deserialize)]
 pub struct Task {
     pub id: Uuid,
     pub description: String,
